# Refactor workflow setup to improve performance